backend/requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements/test.txt requirements/test.in
#
aiohappyeyeballs==2.6.1
    # via
    #   -r requirements/base.txt
    #   aiohttp
aiohttp==3.13.0
    # via
    #   -r requirements/base.txt
    #   litellm
aiosignal==1.4.0
    # via
    #   -r requirements/base.txt
    #   aiohttp
annotated-types==0.7.0
    # via
    #   -r requirements/base.txt
    #   pydantic
anyio==4.11.0
    # via
    #   -r requirements/base.txt
    #   httpx
    #   openai
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
attrs==25.4.0
    # via
    #   -r requirements/base.txt
    #   aiohttp
    #   jsonschema
    #   referencing
certifi==2025.10.5
    # via
    #   -r requirements/base.txt
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via pynacl
charset-normalizer==3.4.4
    # via
    #   -r requirements/base.txt
    #   requests
click==8.3.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   litellm
code-annotations==1.8.2
    # via -r requirements/test.in
coverage[toml]==7.8.0
    # via pytest-cov
distro==1.9.0
    # via
    #   -r requirements/base.txt
    #   openai
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-extensions
    #   django-filter
    #   django-model-utils
    #   django-waffle
    #   djangorestframework
    #   edx-django-release-util
    #   edx-django-utils
    #   edx-submissions
    #   jsonfield
django-crum==0.7.9
    # via edx-django-utils
django-extensions==4.1
    # via -r requirements/test.in
django-filter==25.1
    # via -r requirements/base.txt
django-model-utils==5.0.0
    # via
    #   -r requirements/base.txt
    #   edx-submissions
django-waffle==4.2.0
    # via edx-django-utils
djangorestframework==3.16.0
    # via
    #   -r requirements/base.txt
    #   edx-submissions
dnspython==2.7.0
    # via
    #   -r requirements/base.txt
    #   pymongo
edx-django-release-util==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-submissions
edx-django-utils==7.4.0
    # via -r requirements/test.in
edx-opaque-keys==2.12.0
    # via -r requirements/base.txt
edx-submissions==3.12.1
    # via -r requirements/base.txt
fastuuid==0.13.5
    # via
    #   -r requirements/base.txt
    #   litellm
filelock==3.20.0
    # via
    #   -r requirements/base.txt
    #   huggingface-hub
frozenlist==1.8.0
    # via
    #   -r requirements/base.txt
    #   aiohttp
    #   aiosignal
fsspec==2025.9.0
    # via
    #   -r requirements/base.txt
    #   huggingface-hub
h11==0.16.0
    # via
    #   -r requirements/base.txt
    #   httpcore
hf-xet==1.1.10
    # via
    #   -r requirements/base.txt
    #   huggingface-hub
httpcore==1.0.9
    # via
    #   -r requirements/base.txt
    #   httpx
httpx==0.28.1
    # via
    #   -r requirements/base.txt
    #   litellm
    #   openai
huggingface-hub==0.35.3
    # via
    #   -r requirements/base.txt
    #   tokenizers
idna==3.11
    # via
    #   -r requirements/base.txt
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via
    #   -r requirements/base.txt
    #   litellm
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   litellm
jiter==0.11.0
    # via
    #   -r requirements/base.txt
    #   openai
jsonfield==3.2.0
    # via
    #   -r requirements/base.txt
    #   edx-submissions
jsonschema==4.25.1
    # via
    #   -r requirements/base.txt
    #   litellm
jsonschema-specifications==2025.9.1
    # via
    #   -r requirements/base.txt
    #   jsonschema
litellm==1.78.2
    # via -r requirements/base.txt
markupsafe==3.0.3
    # via
    #   -r requirements/base.txt
    #   jinja2
multidict==6.7.0
    # via
    #   -r requirements/base.txt
    #   aiohttp
    #   yarl
newrelic==10.9.0
    # via edx-django-utils
openai==2.4.0
    # via
    #   -r requirements/base.txt
    #   litellm
openedx-atlas==0.7.0
    # via -r requirements/base.txt
packaging==25.0
    # via
    #   -r requirements/base.txt
    #   huggingface-hub
    #   pytest
pbr==6.1.1
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.5.0
    # via pytest
propcache==0.4.1
    # via
    #   -r requirements/base.txt
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via edx-django-utils
pycparser==2.22
    # via cffi
pydantic==2.11.10
    # via
    #   -r requirements/base.txt
    #   litellm
    #   openai
pydantic-core==2.33.2
    # via
    #   -r requirements/base.txt
    #   pydantic
pymongo==4.12.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via edx-django-utils
pytest==8.3.5
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.1.1
    # via -r requirements/test.in
pytest-django==4.11.1
    # via -r requirements/test.in
python-dotenv==1.1.1
    # via
    #   -r requirements/base.txt
    #   litellm
python-slugify==8.0.4
    # via code-annotations
pytz==2025.2
    # via
    #   -r requirements/base.txt
    #   edx-submissions
pyyaml==6.0.3
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-release-util
    #   huggingface-hub
referencing==0.37.0
    # via
    #   -r requirements/base.txt
    #   jsonschema
    #   jsonschema-specifications
regex==2025.9.18
    # via
    #   -r requirements/base.txt
    #   tiktoken
requests==2.32.5
    # via
    #   -r requirements/base.txt
    #   huggingface-hub
    #   tiktoken
rpds-py==0.27.1
    # via
    #   -r requirements/base.txt
    #   jsonschema
    #   referencing
six==1.17.0
    # via
    #   -r requirements/base.txt
    #   edx-django-release-util
sniffio==1.3.1
    # via
    #   -r requirements/base.txt
    #   anyio
    #   openai
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.4.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
tiktoken==0.12.0
    # via
    #   -r requirements/base.txt
    #   litellm
tokenizers==0.22.1
    # via
    #   -r requirements/base.txt
    #   litellm
tqdm==4.67.1
    # via
    #   -r requirements/base.txt
    #   huggingface-hub
    #   openai
typing-extensions==4.13.2
    # via
    #   -r requirements/base.txt
    #   aiosignal
    #   anyio
    #   edx-opaque-keys
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   -r requirements/base.txt
    #   pydantic
urllib3==2.5.0
    # via
    #   -r requirements/base.txt
    #   requests
yarl==1.22.0
    # via
    #   -r requirements/base.txt
    #   aiohttp
zipp==3.23.0
    # via
    #   -r requirements/base.txt
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
